Fix small problems with decimals and update docker compose

+++ src/RapidPay.Domain/Services/IPaymentFeesModule.cs
@@ -2,5 +2,5 @@ namespace RapidPay.Domain.Services;
 
 public interface IPaymentFeesModule
 {
-    double GetFee();
+    decimal GetFee();
 }
